requirements.txt
aiohttp[speedups]==3.7.4
psutil==5.8.0
pyjwt==2.0.1
tenacity==6.3.1

# Sanic
sanic==20.12.2

# GraphQL
graphene==2.1.8
graphql-server-core==1.1.0
sanic-graphql
aiodataloader==0.2.0

# When graphene 3 is available
# graphql-server[sanic]

# Database
sqlalchemy==1.4.0b3
asyncpg==0.22.0

# Database migration
alembic==1.5.5
psycopg2-binary==2.8.6

# Cache
aioredis==1.3.1

# Instrument
sentry-sdk
